Refactor lib tests into own file (#18)
